requirements.txt
scrapingbee
Pillow
requests
python-dotenv
pillow-heif
pillow-avif-plugin
pyjxl